tratamento de exceções customizado

+++ src/main/java/com/bruel/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.bruel.dscommerce.services;
 import com.bruel.dscommerce.dtos.ProductDTO;
 import com.bruel.dscommerce.entities.Product;
 import com.bruel.dscommerce.repositories.ProductRepository;
+import com.bruel.dscommerce.services.exceptions.DatabaseException;
+import com.bruel.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,7 +25,8 @@ public class ProductService {
     //o service retorna um DTO
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product result = productRepository.findById(id).get();
+        Product result = productRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(result);
     }
 
@@ -44,18 +50,32 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO){
-        //prepara um objeto monitorado pelo JPA
-        Product product = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO, product);
-        //salva a requisição no banco e obtem uma nova referência para ele
-        product = productRepository.save(product);
+        try{
+            //prepara um objeto monitorado pelo JPA
+            Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, product);
+            //salva a requisição no banco e obtem uma nova referência para ele
+            product = productRepository.save(product);
 
-        return new ProductDTO(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
-    @Transactional
+    //executa a transação se o metódo estiver no contexto de outra transação
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        productRepository.deleteById(id);
+        if(!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try{
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product entity){
